+++ src/timer.rs
@@ -1,6 +1,3 @@
-#[cfg(feature = "singleton")]
-use crate::util::AmoOnceRef;
-
 /// Timer programmer support
 pub trait Timer: Send + Sync {
     /// Programs the clock for next event after `stime_value` time.
@@ -17,33 +14,5 @@ impl<T: Timer> Timer for &T {
     #[inline]
     fn set_timer(&self, stime_value: u64) {
         T::set_timer(self, stime_value)
-    }
-}
-
-#[cfg(feature = "singleton")]
-static TIMER: AmoOnceRef<dyn Timer> = AmoOnceRef::new();
-
-#[cfg(feature = "singleton")]
-/// Init TIMER module
-pub fn init_timer(timer: &'static dyn Timer) {
-    if !TIMER.try_call_once(timer) {
-        panic!("load sbi module when already loaded")
-    }
-}
-
-#[cfg(feature = "singleton")]
-#[inline]
-pub(crate) fn probe_timer() -> bool {
-    TIMER.get().is_some()
-}
-
-#[cfg(feature = "singleton")]
-#[inline]
-pub(crate) fn set_timer(time_value: u64) -> bool {
-    if let Some(timer) = TIMER.get() {
-        timer.set_timer(time_value);
-        true
-    } else {
-        false
     }
 }